% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dk_summarize_with_totals.R
\name{dk_means_by_groups_and_total}
\alias{dk_means_by_groups_and_total}
\title{Title}
\usage{
dk_means_by_groups_and_total(.data, group_by_var, mean_var)
}
\arguments{
\item{mean_var}{}
}
\description{
Title
}
