% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{ggsave_pixels}
\alias{ggsave_pixels}
\title{Save ggplot using pixels for dimensions}
\usage{
ggsave_pixels(plot, width = 640, height = 1280, text_factor = 1, filename)
}
\arguments{
\item{plot}{Name of plot to save}

\item{width}{Width in pixels}

\item{height}{Height in pixels}

\item{filename}{Filename to save to}
}
\value{

}
\description{
Save ggplot using pixels for dimensions
}
