% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkmisc-package.R
\docType{package}
\name{dkmisc-package}
\alias{dkmisc}
\alias{dkmisc-package}
\title{dkmisc: David Keyes Personal R Package}
\description{
David Keyes Personal R Package
}
\author{
\strong{Maintainer}: David Keyes \email{hello@dgkeyes.com}

}
\keyword{internal}
